SONARJAVA-1091 Handle report format for coverage per test
